Move genesis/common block functionality into their own file
